# spec/openapi.yaml
openapi: 3.0.0
info:
  title: Календарь звонков
  description: |-
    «Календарь звонков» — упрощённый сервис бронирования времени по мотивам Cal.com.

    Роли (регистрации и авторизации в приложении нет):
    - Владелец календаря — один заранее заданный профиль, работает в админской части
    под префиксом `/admin`. Создаёт типы событий и смотрит список предстоящих встреч.
    - Гость — без аккаунта. Смотрит виды брони, выбирает свободный слот и записывается.

    Договорённости, общие для всего контракта:
    - Всё время — UTC, формат ISO 8601 (`2026-08-20T09:00:00Z`). Таймзон в приложении нет.
    - Идентификаторы — непрозрачные строки (UUID). Клиент не разбирает их и не считает
    порядковыми.
    - Окно записи — 14 дней от текущего момента.
    - Слоты нарезаются по сетке стартов с шагом 30 минут, в рабочие часы 09:00–18:00 UTC,
    только по будням (пн–пт).
    - Бронь занимает время `[start, start + durationMinutes)` выбранного типа события.
    Событие длиннее 30 минут блокирует несколько соседних стартов.
    - Занятость — глобальная: на пересекающееся время нельзя создать две записи,
    даже если это разные типы событий.
    - Любая ошибка возвращается телом `{ error, message }`: `error` — машиночитаемый код
    для клиента, `message` — человекочитаемое пояснение.
    - Ошибки, общие для всех путей и потому не описанные в каждой операции отдельно:
    `404 not_found` — пути нет в контракте (а не HTML-страница Express);
    `400 validation_failed` — тело не разобралось как JSON;
    `413 payload_too_large` — тело больше 32 КБ;
    `415 unsupported_media_type` — неподдерживаемая кодировка тела;
    `500 internal_error` — всё остальное. Форма тела у всех одна.
  version: 0.0.0
tags:
  - name: Гость
  - name: Владелец
paths:
  /admin/bookings:
    get:
      operationId: AdminBookings_list
      description: |-
        Предстоящие встречи одним списком по всем типам событий,
        по возрастанию времени начала. Прошедшие встречи не возвращаются.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
      tags:
        - Владелец
  /admin/event-types:
    get:
      operationId: AdminEventTypes_list
      description: Все типы событий владельца.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventType'
      tags:
        - Владелец
    post:
      operationId: AdminEventTypes_create
      description: Создать тип события.
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        '400':
          description: 'Тело запроса не прошло валидацию: не те типы, не та сетка, время вне окна записи.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
      tags:
        - Владелец
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventTypeCreate'
  /bookings:
    post:
      operationId: PublicBookings_create
      description: Записаться на свободный слот.
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: 'Тело запроса не прошло валидацию: не те типы, не та сетка, время вне окна записи.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '404':
          description: Типа события с таким `id` не существует.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventTypeNotFound'
        '409':
          description: |-
            Слот уже занят: запрошенный интервал пересекается с существующей бронью
            (в том числе бронью другого типа события).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlotUnavailable'
      tags:
        - Гость
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreate'
  /event-types:
    get:
      operationId: PublicEventTypes_list
      description: Список видов брони на публичной странице.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventType'
      tags:
        - Гость
  /event-types/{id}:
    get:
      operationId: PublicEventTypes_read
      description: Один тип события — для прямой ссылки на страницу бронирования.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        '404':
          description: Типа события с таким `id` не существует.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventTypeNotFound'
      tags:
        - Гость
  /event-types/{id}/slots:
    get:
      operationId: PublicEventTypes_slots
      description: |-
        Свободные слоты выбранного типа события на 14 дней вперёд,
        по возрастанию времени. Занятые интервалы уже исключены.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Slot'
        '404':
          description: Типа события с таким `id` не существует.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventTypeNotFound'
      tags:
        - Гость
components:
  schemas:
    Booking:
      type: object
      required:
        - id
        - eventTypeId
        - eventTypeTitle
        - start
        - end
        - guestName
        - guestEmail
        - createdAt
      properties:
        id:
          type: string
          description: Непрозрачный идентификатор (UUID).
        eventTypeId:
          type: string
        eventTypeTitle:
          type: string
          minLength: 1
          maxLength: 100
          description: Название типа события на момент чтения — чтобы список владельца не требовал второго запроса.
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        guestName:
          type: string
          minLength: 1
          maxLength: 100
        guestEmail:
          type: string
          format: email
        notes:
          type: string
          description: Необязательный комментарий гостя.
        createdAt:
          type: string
          format: date-time
      description: Запись гостя на звонок.
    BookingCreate:
      type: object
      required:
        - eventTypeId
        - start
        - guestName
        - guestEmail
      properties:
        eventTypeId:
          type: string
        start:
          type: string
          format: date-time
          description: Начало выбранного слота; должно совпадать со `start` одного из свободных слотов.
        guestName:
          type: string
          minLength: 1
          maxLength: 100
        guestEmail:
          type: string
          format: email
        notes:
          type: string
          maxLength: 1000
      description: Тело запроса на создание брони.
    EventType:
      type: object
      required:
        - id
        - title
        - description
        - durationMinutes
      properties:
        id:
          type: string
          description: Непрозрачный идентификатор (UUID).
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: Название, которое видит гость в списке видов брони.
        description:
          type: string
          maxLength: 1000
          description: Описание встречи для страницы бронирования.
        durationMinutes:
          type: integer
          format: int32
          minimum: 30
          maximum: 240
          description: Длительность встречи в минутах; кратна 30.
      description: Тип встречи, который публикует владелец календаря.
    EventTypeCreate:
      type: object
      required:
        - title
        - description
        - durationMinutes
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 1000
        durationMinutes:
          type: integer
          format: int32
          minimum: 30
          maximum: 240
          description: 'Кратна 30: 30, 60, 90, 120.'
      description: Тело запроса на создание типа события.
    EventTypeNotFound:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - event_type_not_found
        message:
          type: string
      description: Типа события с таким `id` не существует.
    Slot:
      type: object
      required:
        - start
        - end
      properties:
        start:
          type: string
          format: date-time
          description: Начало встречи, всегда попадает на сетку с шагом 30 минут.
        end:
          type: string
          format: date-time
          description: 'Конец встречи: `start + durationMinutes` типа события.'
      description: |-
        Свободный интервал под конкретный тип события.
        Слоты не хранятся — они вычисляются на лету из окна записи и существующих броней.
    SlotUnavailable:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - slot_unavailable
        message:
          type: string
      description: |-
        Слот уже занят: запрошенный интервал пересекается с существующей бронью
        (в том числе бронью другого типа события).
    ValidationError:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - validation_failed
        message:
          type: string
      description: 'Тело запроса не прошло валидацию: не те типы, не та сетка, время вне окна записи.'
servers:
  - url: http://localhost:3000
    description: Локальный запуск
    variables: {}